authentication

+++ app/Http/Controllers/AuthorController.php
@@ -14,10 +14,11 @@ class AuthorController extends Controller
      */
     public function index()
     {
+        $author = Author::get();
         return response()->json([
-            "message" => "Daftar Author",
-            "data" => Author::all()
-        ]);
+            'status' => 200,
+            'data' => $author
+        ], 200);
     }
 
     /**
@@ -46,7 +47,19 @@ class AuthorController extends Controller
         $author->email = $request->email;
         $author->hp = $request->hp;
         $author->save();
-        return response()->json($author, 202); 
+        if($author->save()){
+            return response()->json([
+                'status' => 201,
+                'message' => 'Data Berhasil Di Simpan!',
+                'data' => $author
+            ], 201);
+        }
+        else{
+            return response()->json([
+                'status' => 400,
+                'message' => 'Data Belum Tersimpan!'
+            ], 400);
+        }
     }
 
     /**
@@ -58,7 +71,19 @@ class AuthorController extends Controller
     public function show($id)
     {
         $author = Author::find($id);
-        return response()->json($author);
+        if($author){
+            return response()->json([
+                'status' => 200,
+                'message' => 'Data berhasil ditemukan!',
+                'data' => $author
+            ], 200);
+        }
+        else{
+            return response()->json([
+                'status' => 404,
+                'message' => 'Id ' . $id . ' tidak ditemukan!'
+            ], 404);
+        }
     }
 
     /**
@@ -82,14 +107,26 @@ class AuthorController extends Controller
     public function update(Request $request, $id)
     {
         $author = Author::find($id);
-        $author->name = $request->name;
-        $author->date_of_birth = $request->date_of_birth;
-        $author->place_of_birth = $request->place_of_birth;
-        $author->gender = $request->gender;
-        $author->email = $request->email;
-        $author->hp = $request->hp;
-        $author->save();
-        return response()->json($author, 202);
+        if($author){
+            $author->name = $request->name ? $request->name : $author->name;
+            $author->date_of_birth = $request->date_of_birth ? $request->date_of_birth : $author->date_of_birth;
+            $author->place_of_birth = $request->place_of_birth ? $request->place_of_birth : $author->place_of_birth;
+            $author->gender = $request->gender ? $request->gender : $author->gender;
+            $author->email = $request->email ? $request->email : $author->email;
+            $author->hp = $request->hp ? $request->hp : $author->hp;
+            $author->save();
+            return response()->json([
+                'status' => 200,
+                'message' => 'Data berhasil diubah!',
+                'data' => $author
+            ], 200);
+        }
+        else{
+            return response()->json([
+                'status' => 404,
+                'message' => 'Id ' . $id . ' tidak ditemukan!'
+            ], 404);
+        }
     }
 
     /**
@@ -100,11 +137,19 @@ class AuthorController extends Controller
      */
     public function destroy($id)
     {
-        $author = Author::findOrFail($id);
-        $author->delete();
-        return response()->json([
-            "message" => "Data Author Telah dihapus",
-            "data" => $author
-        ]);
+        $author = Author::find($id);
+        if($author){
+            $author->delete();
+            return response()->json([
+                'status' => 200,
+                'message' => 'Data Id ' . $id . ' Berhasil dihapus!'
+            ], 200);
+        }
+        else{
+            return response()->json([
+                'status' => 404,
+                'message' => 'Id ' . $id . ' tidak ditemukan!'
+            ], 404);
+        }
     }
 }
